refactor: api moved to pkg
