+++ pkg/task/append/append.go
@@ -7,7 +7,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/sn3d/excav/api"
+	"github.com/sn3d/excav/pkg/api"
 	"github.com/sn3d/excav/pkg/cast"
 	"github.com/sn3d/excav/pkg/log"
 	"github.com/sn3d/excav/pkg/template"
